requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==2.1.1
constantly==15.1.0
crispy-bootstrap5==0.7
cryptography==40.0.2
daphne==4.0.0
Deprecated==1.2.13
dj-database-url==1.2.0
Django==4.1.3
django-braces==1.15.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.6.0
django-embed-video==1.4.4
django-extensions==3.2.1
django-filter==23.1
django-model-utils==4.3.1
django-private-chat2==1.0.2
django-redisboard==8.3.0
djangorestframework==3.14.0
hiredis==2.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
packaging==23.1
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pymemcache==3.5.2
pyOpenSSL==23.1.1
pytz==2022.7
redis==4.5.5
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
websockets==8.1
wrapt==1.15.0
zope.interface==6.0
